Made username, password, email not nullable so that they cannot be empty inside of the form. Also added persistence tests and application tests.
