+++ auth-service/src/main/java/com/petclinic/auth/User/User.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import javax.validation.constraints.NotNull;
 import java.util.List;
 import java.util.Set;
 
@@ -22,8 +23,13 @@ public class User {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
+    @NotNull
     private String username;
+
+    @NotNull
     private String password;
+
+    @NotNull
     private String email;
 
     public User(String username, String password, String email) {
